backend/requirements_agent.txt
# LiveKit Agents Framework - Dépendances pour l'agent moderne
livekit-agents>=0.8.0
livekit-plugins-silero>=0.6.0
livekit-api>=0.5.0
livekit>=0.10.0

# HTTP et réseau
httpx>=0.25.0
aiohttp>=3.8.0

# Audio et traitement
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0

# Utilitaires
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.13.0

# Logging et monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Développement et debug
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0